+++ mkdocs.yml
@@ -46,7 +46,6 @@ theme:
   features:
     - navigation.sections
     - navigation.indexes
-    - navigation.expand
     - navigation.path
     - navigation.instant
     - navigation.instant.prefetch
@@ -99,6 +98,7 @@ plugins:
             separate_signature: true
             signature_crossrefs: true
             show_if_no_docstring: false # ignore undocumented members
+            show_overloads: false
 
             show_symbol_type_toc: true
             show_root_full_path: true # prairielearn.core
